+++ json_serializable/test/test_files/json_test_example.wrapped.g.dart
@@ -13,10 +13,7 @@ part of 'json_test_example.wrapped.dart';
 Person _$PersonFromJson(Map<String, dynamic> json) => new Person(
     json['firstName'] as String,
     json['lastName'] as String,
-    json[r'$house'] == null
-        ? null
-        : House.values
-            .singleWhere((e) => e.toString() == 'House.${json[r'$house']}'),
+    $enumDecodeNullable('House', House.values, json[r'$house'] as String),
     middleName: json['middleName'] as String,
     dateOfBirth: json['dateOfBirth'] == null
         ? null
@@ -25,11 +22,7 @@ Person _$PersonFromJson(Map<String, dynamic> json) => new Person(
       ? null
       : new Order.fromJson(json['order'] as Map<String, dynamic>)
   ..houseMap = (json['houseMap'] as Map<String, dynamic>)?.map((k, e) =>
-      new MapEntry(
-          k,
-          e == null
-              ? null
-              : House.values.singleWhere((e2) => e2.toString() == 'House.$e')));
+      new MapEntry(k, $enumDecodeNullable('House', House.values, e as String)));
 
 abstract class _$PersonSerializerMixin {
   String get firstName;
@@ -83,8 +76,7 @@ class _$PersonJsonMapWrapper extends $JsonMapWrapper {
 }
 
 Order _$OrderFromJson(Map<String, dynamic> json) => new Order(
-    Category.values
-        .singleWhere((e) => e.toString() == 'Category.${json['category']}'),
+    $enumDecode('Category', Category.values, json['category'] as String),
     (json['items'] as List)?.map(
         (e) => e == null ? null : new Item.fromJson(e as Map<String, dynamic>)))
   ..count = json['count'] as int