+++ json_annotation/lib/src/json_serializable.dart
@@ -4,6 +4,13 @@
 
 /// An annotation used to specify a class to generate code for.
 class JsonSerializable {
+  /// If `true` (the default), then any unrecognized keys passed to the
+  /// generated FromJson factory will be ignored.
+  ///
+  /// If `false`, any unrecognized keys will be treated as an error.
+  /// ```
+  final bool allowUnrecognizedKeys;
+
   /// If `true` (the default), a private, static `_$ExampleFromJson` method
   /// is created in the generated part file.
   ///
@@ -48,11 +55,13 @@ class JsonSerializable {
 
   /// Creates a new [JsonSerializable] instance.
   const JsonSerializable(
-      {bool createFactory: true,
+      {bool allowUnrecognizedKeys: true,
+      bool createFactory: true,
       bool createToJson: true,
       bool includeIfNull: true,
       bool nullable: true})
-      : this.createFactory = createFactory ?? true,
+      : this.allowUnrecognizedKeys = allowUnrecognizedKeys ?? true,
+        this.createFactory = createFactory ?? true,
         this.createToJson = createToJson ?? true,
         this.includeIfNull = includeIfNull ?? true,
         this.nullable = nullable ?? true;